backend/requirements.txt
fastapi==0.116.1
protobuf==6.32.0
python-dotenv==1.1.1
